requirements-minimal.txt
# Minimal requirements for web app testing
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.0
python-multipart>=0.0.6
pydantic>=2.4.0
pandas>=2.1.0
pyyaml>=6.0
python-dotenv>=1.0.0
rich>=13.0.0
typer>=0.9.0
tqdm>=4.66.0